move plotting.surf_plotting under plotting.surface
