+++ nilearn/plotting/surface/__init__.py
@@ -1,0 +1,15 @@
+from nilearn.plotting.surface.surf_plotting import (
+    plot_img_on_surf,
+    plot_surf,
+    plot_surf_contours,
+    plot_surf_roi,
+    plot_surf_stat_map,
+)
+
+__all__ = [
+    "plot_img_on_surf",
+    "plot_surf",
+    "plot_surf_contours",
+    "plot_surf_roi",
+    "plot_surf_stat_map",
+]
